Уборка, конфигурация сущностей, UserService и долги

+++ backend/Money.Api/Dto/Categories/CategoryDto.cs
@@ -49,7 +49,7 @@ public class CategoryDto
             ParentId = category.ParentId,
             Order = category.Order,
             Color = category.Color,
-            OperationTypeId = category.OperationType,
+            OperationTypeId = (int)category.OperationType,
         };
     }
 }
